+++ app/controllers/supporting_information_files_controller.rb
@@ -1,31 +1,35 @@
 class SupportingInformationFilesController < ApplicationController
   respond_to :json
   before_action :authenticate_user!
-  before_action :enforce_policy
 
   def show
+    requires_user_can(:view, file.paper)
     respond_with file
   end
 
   def create
-    file.update_attributes(status: 'processing')
-    DownloadAttachmentWorker.perform_async(file.id, params[:url], current_user.id)
+    requires_user_can(:edit, file.paper)
+    DownloadAttachmentWorker
+      .download_attachment(file, params[:url], current_user)
     respond_with file
   end
 
   def update
+    requires_user_can(:edit, file.paper)
     file.update_attributes file_params
     respond_with file
   end
 
   def destroy
+    requires_user_can(:edit, file.paper)
     file.destroy
     head :no_content
   end
 
   def update_attachment
-    file.update_attribute(:status, 'processing')
-    DownloadAttachmentWorker.perform_async(file.id, params[:url], current_user.id)
+    requires_user_can(:edit, file.paper)
+    DownloadAttachmentWorker
+      .download_attachment(file, params[:url], current_user)
     respond_with file
   end
 
@@ -46,10 +50,6 @@ class SupportingInformationFilesController < ApplicationController
     task.supporting_information_files.new(paper_id: task.paper_id)
   end
 
-  def enforce_policy
-    authorize_action!(resource: file)
-  end
-
   def file_params
     params.require(:supporting_information_file).permit(
       :title,
@@ -62,7 +62,4 @@ class SupportingInformationFilesController < ApplicationController
       attachment: [])
   end
 
-  def render_404
-    head 404
-  end
 end